+++ stub.php
@@ -1,5 +1,6 @@
 <?php
 Phar::mapPhar('connector.phar');
+Phar::interceptFileFuncs();
 
 include_once 'phar://connector.phar/vendor/autoload.php';